refactor: reorganize appr details structure
