+++ web/web.njsproj
@@ -66,7 +66,6 @@
    <Content Include="src\assets\calendar-select.svg" />
     <Content Include="src\assets\video.svg" />
     <Content Include="src\components\civil\CivilAdjudicatorRestrictions.vue" />
-    <Content Include="src\components\civil\CivilAppearanceDetails.vue" />
     <Content Include="src\components\civil\CivilCaseDetails.vue" />
     <Content Include="src\components\civil\CivilDocumentsView.vue" />
     <Content Include="src\components\civil\CivilFutureAppearances.vue" />
